fix bugs

+++ HeFeiSaledServices/SaledServices/CustomsExport/StockInOutForm.cs
@@ -40,29 +40,39 @@ namespace SaledServices.CustomsExport
         {
             if (isAuto)
             {
-                try
+                string strFilePath = @"D:\logfile\log.txt";
+                if (!File.Exists(strFilePath))
                 {
-                    string strFilePath = @"D:\logfile\log.txt";
-                    string strOldText = File.ReadAllText(strFilePath, System.Text.Encoding.Default);
+                    Directory.CreateDirectory(@"D:\logfile");
+                    File.Create(strFilePath);
+                }
 
-                    //需要发邮件通知
-                    //SendMail("qxmin1984@126.com", "自动对接出错", "qxmin1984@126.com", "认真检查出错信息", info, "qxmin1984", "xmzx19850325", "smtp.126.com", ""); 
-                    if (isError)
+                if (File.Exists(strFilePath))
+                {
+                    try
                     {
-                        SendMail("xinmin.qi@maiweibao.com.cn", "海关自动对接出错", "qxmin1984@126.com", "认真检查出错信息", info, "xinmin.qi@maiweibao.com.cn", "Mwb20180802", "smtp.mxhichina.com", 25);
-                        info = "ERROR: " + info + "   " + DateTime.Now.ToString("O") + "\r\n" + strOldText;
-                    }
+                        string strOldText = File.ReadAllText(strFilePath, System.Text.Encoding.Default);
 
-                    File.WriteAllText(strFilePath, info, System.Text.Encoding.Default);
+                        //需要发邮件通知
+                        //SendMail("qxmin1984@126.com", "自动对接出错", "qxmin1984@126.com", "认真检查出错信息", info, "qxmin1984", "xmzx19850325", "smtp.126.com", ""); 
+                        if (isError)
+                        {
+                            SendMail("xinmin.qi@maiweibao.com.cn", "海关自动对接出错", "qxmin1984@126.com", "认真检查出错信息", info, "xinmin.qi@maiweibao.com.cn", "Mwb20180802", "smtp.mxhichina.com", 25);
+                            info = "ERROR: " + info + "   " + DateTime.Now.ToString("O") + "\r\n" + strOldText;
+                        }
+
+                        File.WriteAllText(strFilePath, info, System.Text.Encoding.Default);
+
+                    }
+                    catch (Exception ex)
+                    {
+                        SendMail("xinmin.qi@maiweibao.com.cn", "日志记录出错", "qxmin1984@126.com", "认真检查出错信息", ex.ToString(), "xinmin.qi@maiweibao.com.cn", "Mwb20180802", "smtp.mxhichina.com", 25);
+                    }
                 }
-                catch (Exception ex)
+                else
                 {
-                    SendMail("xinmin.qi@maiweibao.com.cn", "日志记录出错", "qxmin1984@126.com", "认真检查出错信息", ex.ToString(), "xinmin.qi@maiweibao.com.cn", "Mwb20180802", "smtp.mxhichina.com", 25);
+                    MessageBox.Show(info);
                 }
-            }
-            else
-            {
-                MessageBox.Show(info);
             }
         }
